🐛 :fix correção rota /chamado DYNAMIC_SERVER_USAGE fronteend
[#146] correção rota /chamado

https://jira.arinterface.com.br/work_packages/146
