+++ .vscode/settings.json
@@ -362,8 +362,10 @@
     "Txid",
     "ufpolo",
     "UNICA",
+    "Unicas",
     "unico",
     "Unico",
+    "Unicos",
     "Updata",
     "urlnota",
     "Userconstrutora",
